+++ metadata.rb
@@ -1,7 +1,7 @@
-name             'slurm-cookbook'
-maintainer       'YOUR_COMPANY_NAME'
-maintainer_email 'YOUR_EMAIL'
+name             'slurm'
+maintainer       'PNNL'
+maintainer_email 'david.brown@pnnl.gov'
 license          'All rights reserved'
-description      'Installs/Configures slurm-cookbook'
+description      'Installs/Configures slurm'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'